+++ src/hooks/useFilterInputChangeHandler.ts
@@ -2,38 +2,11 @@ import { useState } from "react";
 import debounce from "lodash.debounce";
 import queryString from "query-string";
 import { useLocation, useHistory } from "react-router-dom";
-import { parseQueryString, stringifyQuery } from "utils";
+import { updateUrlQueryParam } from "utils/url";
 
 const arrayFormat = "comma";
 
-const updateQueryParam = (
-  urlSearchParam: string,
-  inputValue: string | string[] | null,
-  search: string,
-  replace: (path: string) => void,
-  pathname: string,
-  sendAnalyticsEvent: (filterBy: string) => void,
-  resetPage?: boolean
-) => {
-  const urlParams = parseQueryString(search);
-
-  if (!inputValue) {
-    delete urlParams[urlSearchParam];
-  } else {
-    urlParams[urlSearchParam] = inputValue;
-  }
-
-  const nextQueryParams = stringifyQuery({
-    ...urlParams,
-    ...(resetPage && { page: 0 }),
-  });
-
-  replace(`${pathname}?${nextQueryParams}`);
-
-  sendAnalyticsEvent(urlSearchParam);
-};
-
-const updateQueryParamWithDebounce = debounce(updateQueryParam, 250);
+const updateQueryParamWithDebounce = debounce(updateUrlQueryParam, 250);
 
 type InputEvent = React.ChangeEvent<HTMLInputElement>;
 
@@ -75,122 +48,4 @@ export const useFilterInputChangeHandler = (
   };
 
   return [value, onChange];
-};
-
-interface Params<SearchParam> {
-  urlSearchParam: SearchParam; // the name of url param
-  sendAnalyticsEvent: (filterBy: string) => void; // callback function to send analytics event
-}
-
-type UseInputFilterReturn = [
-  string, // url param value
-  (e: InputEvent) => void, // onChange handler
-  () => void, // update url param
-  () => void // reset url param
-];
-
-// USE FOR FILTERS BUILT INTO TABLE COLUMN HEADERS
-export const useInputFilter = <SearchParam extends string>({
-  urlSearchParam,
-  sendAnalyticsEvent = () => undefined,
-}: Params<SearchParam>): UseInputFilterReturn => {
-  const { pathname, search } = useLocation();
-  const { replace } = useHistory();
-
-  const parsed = queryString.parse(search, { arrayFormat });
-
-  const inputValueFromUrl = (parsed[urlSearchParam] || "").toString();
-
-  const [value, setValue] = useState(inputValueFromUrl);
-
-  const onChange = (e: InputEvent): void => {
-    setValue(e.target.value);
-  };
-
-  const updateUrlSearchParam = () => {
-    updateQueryParam(
-      urlSearchParam,
-      value,
-      search,
-      replace,
-      pathname,
-      sendAnalyticsEvent,
-      true // when does this need to be false?
-    );
-
-    sendAnalyticsEvent(urlSearchParam);
-  };
-
-  const resetQueryParam = () => {
-    setValue("");
-
-    updateQueryParam(
-      urlSearchParam,
-      null,
-      search,
-      replace,
-      pathname,
-      sendAnalyticsEvent,
-      true // when does this need to be false?
-    );
-  };
-
-  return [value, onChange, updateUrlSearchParam, resetQueryParam];
-};
-
-type UseTreeSelectFilterReturn = [
-  string[], // url param value
-  (e: string[]) => void, // onChange handler
-  () => void, // update url param
-  () => void // reset url param
-];
-
-export const useTreeSelectFilter = <SearchParam extends string>({
-  urlSearchParam,
-  sendAnalyticsEvent = () => undefined,
-}: Params<SearchParam>): UseTreeSelectFilterReturn => {
-  const { pathname, search } = useLocation();
-  const { replace } = useHistory();
-
-  const { [urlSearchParam]: rawStatuses } = parseQueryString(search);
-
-  const valueFromUrl = Array.isArray(rawStatuses)
-    ? rawStatuses
-    : [rawStatuses].filter((v) => v);
-
-  const [value, setValue] = useState<string[]>(valueFromUrl);
-
-  const onChange = (newValue: string[]): void => {
-    setValue(newValue);
-  };
-
-  const updateUrlSearchParam = () => {
-    updateQueryParam(
-      urlSearchParam,
-      value,
-      search,
-      replace,
-      pathname,
-      sendAnalyticsEvent,
-      true // when does this need to be false?
-    );
-
-    sendAnalyticsEvent(urlSearchParam);
-  };
-
-  const resetQueryParam = () => {
-    setValue([]);
-
-    updateQueryParam(
-      urlSearchParam,
-      null,
-      search,
-      replace,
-      pathname,
-      sendAnalyticsEvent,
-      true // when does this need to be false?
-    );
-  };
-
-  return [value, onChange, updateUrlSearchParam, resetQueryParam];
 };
