+++ src/utils/url/updateUrlQueryParam.ts
@@ -1,0 +1,28 @@
+import { parseQueryString, stringifyQuery } from "utils";
+
+export const updateUrlQueryParam = (
+  urlSearchParam: string,
+  inputValue: string | string[] | null,
+  search: string,
+  replace: (path: string) => void,
+  pathname: string,
+  sendAnalyticsEvent: (filterBy: string) => void,
+  resetPage?: boolean
+) => {
+  const urlParams = parseQueryString(search);
+
+  if (!inputValue) {
+    delete urlParams[urlSearchParam];
+  } else {
+    urlParams[urlSearchParam] = inputValue;
+  }
+
+  const nextQueryParams = stringifyQuery({
+    ...urlParams,
+    ...(resetPage && { page: 0 }),
+  });
+
+  replace(`${pathname}?${nextQueryParams}`);
+
+  sendAnalyticsEvent(urlSearchParam);
+};
